EvseV2G: simplify dlog() wrapper
While not converting the module to standard EVerest EVLOG_*, this
significantly simplifies the dlog() cruft to the most necessary bits.

As a side effect, the annoying empty lines on stdout logging are removed.

This also converts the macro to a function.

Also drop obviously obsolete trailing newlines from dlog() calls.

Signed-off-by: Moritz Barsnick <moritz.barsnick@chargebyte.com>