SerialCommHub: make retry count on error configurable
At the moment, an automatic retry up to 2 times is done in case the
Modbus operation fails. With some devices or in same use-cases, this
is not desired, so let us make the retry count configurable via
manifest.

So mimic the current behavior and to retain backwards compatibility,
let's use a default value of 2.

While at, we slightly change the debug messages and unify the
loop counting code.

Signed-off-by: Michael Heimpold <michael.heimpold@chargebyte.com>