RpcApi: convert type of meter timestamp to string
It is now represented as an RFC3339 string, as also used within EVerest.

The previously used "number" resulted in a C float type, which does not have
enough precision for epoch seconds. (In the current epoch range, it can
only represent the epoch's integer in steps of 128.) C double type is not
available.

Signed-off-by: Moritz Barsnick <moritz.barsnick@chargebyte.com>