RpcApi: Refactor EVSE status handling in check_evse_session_event method
For some reason, the source.enable_state is never reported as Enable, so the
EVSE status is always set to false. It is only sending an Unassigned state,
which has led to an available state of false.

Now we initialize the available state to true in the constructor of
DataStore, so that the EVSE status is always set to true when the EVSE is
enabled.

Signed-off-by: Fabian Hartung <fabian.hartung@chargebyte.com>