EvseV2G: Replace evse_replug method with reinit state handling
Now it is not longer necessary to implmenent evse_replug command in the
BSP layer. Instead, the EvseManager module can now handle a reinitialization
by calling the low layer commands. Currently, the reinit procedure only
supports the CP state E command for the reinit. This will be expanded
with CP state F and X1.

Signed-off-by: Fabian Hartung <fabian.hartung@chargebyte.com>