EvseManager: Fix issue that a HLC session is detected to late
Until now, subscribe_v2g_setup_finished was used to detect whether an HLC session was active. However, this is not the correct signal, as subscribe_v2g_setup_finished is only called shortly before charging (PowerDelivery). Now it uses the SLAC matched signal to detect a start of a HLC session.

Signed-off-by: Fabian Hartung <fabian.hartung@chargebyte.com>